Package Jackson with the app
These changes add the Jackson MBR/MBW provider and explicitly use
them in the application.  This enables the app to use the packaged
version of Jackson to deserialize the JSON for the LocalDate.

I also updated the test.sh script to include the context root,
since I don't specify one in the server config.